requirements.txt
multiping
pillow